requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.5.0
scikit-learn>=1.2.0
tqdm>=4.65.0
nltk>=3.8.0
